requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.6
django-autoslug==1.9.9
django-filter~=23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
flake8==6.1.0
flake8-absolute-import==1.0.0.2
flake8-commas==2.1.0
flake8-isort==6.1.0
flake8-quotes==3.3.2
flake8-type-ignore==0.1.0.post2
flake8-typing-imports==1.15.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
transliterate==1.10.2
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.6
